+++ Driver.lua
@@ -796,9 +796,19 @@ end
 function TargetedSpellsDriver:EncounterPreventsTTSExecution(unit)
 	-- Lothraxion, Nexus-Point Xenas
 	if self.activeEncounterId == 3333 then
+		local level = UnitLevel(unit)
+
+		if level == -1 then
+			return true
+		end
+
+		local nameplateId = tonumber(string.gsub(unit, "nameplate", ""))
+
+		-- print(unit, level, nameplateId, level == -1 and nameplateId > 1)
+
 		-- ignores casts by a boss unit if the nameplateN id is greater than 1
 		-- Divine Guile spawns 12 images, all UnitLevel == -1
-		return UnitLevel(unit) == -1 and tonumber(string.gsub(unit, "nameplate", "")) > 1
+		return level == -1 and nameplateId > 1
 	end
 
 	return false
@@ -836,7 +846,8 @@ function TargetedSpellsDriver:MaybeAnnounceSpell(info)
 
 	self.ttsAnnouncementCache[info.unit] = now
 
-	local voiceId = TargetedSpellsSaved.Settings.Self.TextToSpeechVoice or C_VoiceChat.GetDefaultVoiceID()
+	local voiceId = TargetedSpellsSaved.Settings.Self.TextToSpeechVoice
+		or C_TTSSettings.GetVoiceOptionID(Enum.TtsVoiceType.Standard)
 
 	C_VoiceChat.SpeakText(voiceId, spellName, 2, C_TTSSettings.GetSpeechVolume(), true)
 end