+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 3.2.1
 
+- fix TTS not playing if you hadn't selected a voice
 - Lothraxion will no longer spam Text-To-Speech announcements for Lingering Image casts
 - fix a bug where some spells with inconsistent events such as Leech Veil would persist as bar elements past the end of their cast time